Simplify mystack.h and add mystack-2.h
mystack.h: add element to queue1, make sure that the queue is empty before adding; delete element in queue1, make sure that there is just one element in queue1 before deleting.

mystack-2.h: add the element to a filled queue (add the element to queue1 in the very beginning, when the two queues are both empty); delete the element in a filled queue, make sure there is just one element in the queue before deleting.